Тесты на добавление банка и транзакции

+++ lib/src/main/java/org/company/cardstat/DatabaseHandler.java
@@ -513,7 +513,7 @@ public class DatabaseHandler extends SQLiteOpenHelper {
      * @return
      * @throws DatabaseHandlerException
      */
-    public BankTransaction getBankTransaction(int _id) throws DatabaseHandlerException {
+    public BankTransaction getBankTransaction(long _id) throws DatabaseHandlerException {
 
         SQLiteDatabase database = this.getReadableDatabase();
